moving Receipt class into model_objects.py
removing unused file receipt.py
