--- python/receipt.py
@@ -1,29 +1,0 @@
-from model_objects import ReceiptItem
-
-class Receipt:
-    def __init__(self):
-        self._items = []
-        self._discounts = []
-
-    @property
-    def total_price(self):
-        total = 0
-        for item in self.items:
-            total += item.total_price
-        for discount in self.discounts:
-            total += discount.discount_amount
-        return total
-
-    @property
-    def items(self):
-        return self._items[:]
-
-    @property
-    def discounts(self):
-        return self._discounts[:]
-
-    def add_product(self, product, quantity, price):
-        self._items.append(ReceiptItem(product, quantity, price))
-
-    def add_discount(self, discount):
-        self._discounts.append(discount)
